Save objects in a more straight way
It's simpler to just save the parent and let the children be autosaved
(and autovalidated) since this is how Rails works out of the box. We
just need to make sure the receipts are properly tied to the parent
notification. We get that, again, by using Rails standard constructs
like `association#build` instead of the ad-hoc `build_receipt` method.

The previous behavior caused issues like getting duplicated errors in
the notification object, because of the fact that everytime a receipt
was saved, the parent notification would be saved too and validations
would be (re)run.